refactoring: - refactored Map model
