+++ src/shared/model/map/index.ts
@@ -1,0 +1,1 @@
+export * from './map';
