+++ src/shared/model/map/map.ts
