+++ src/widgets/Map/hooks/index.ts
@@ -1,0 +1,1 @@
+export { default as useMap } from './useMap';
