--- src/widgets/Map/types/index.ts
@@ -1,1 +1,0 @@
-export * from './types.ts';
